Display: Unsigned comparison causes unreachable code
The number of bytes read by read is returned in a ssize_t so that
errors can be propogated through the return code. Previously we
were storing this return value in an unsigned size_t and then checking
this value against if (size < 0) (Which can never be true)

Simple fix here is to change the size storage to a signed ssize_t

Now that the size is signed, the parameter passed to LayProcess
should be unsigned, so we cast there instead. It's more important
to keep the read calls using a signed value here.

Signed-off-by: Kieran Bingham <kieranbingham@gmail.com>