Display: Convert strlen_onscreen to be unsigned like strlen
Currently there are different use cases of both signed and unsigned
char arrays passed into this function.

Upon initial examination of the function, I believe that it is safe
to use an unsigned type, matching the c library strlen declaration,
and then unifying calls so that they do not have to simply cast
the array types.

The only part of the function which uses the actual (signed/unsigned)
type of the char is the call to FromUtf8() which itself converts to a
signed int.

Converting to use signed char* here seems to be a reasonable fix for the
compiler warnings. (rather than casting the warning lines to unsigned)

Signed-off-by: Kieran Bingham <kieranbingham@gmail.com>